hotfix: various bugfixes (#48)
## Summary

- Fixes bug where when loading the model in `src.model.load_model`, the
`learning_rate` for the Adam optimiser was hardcoded to `0.0001`. It now
uses the correct value from the loaded settings.
- Fixes bug where the global variables `model`, `label_mapping`, and
`stored_sequences` were not reset after training a model. If you
selected 'Create New Model' in `welcome_tab.html` a second time, after
having already trained a model, it would use the state for the old
model.
- Fixes bug where when training a model on video + IMU CSV and then
retraining the model on only a new IMU video, it would use the IMU
telemetry from the old video as `settings['imu_path']` was never reset,
triggering the `if 'video_path' in settings and 'imu_path' in settings:`
check instead of `if 'video_path' in settings and not 'imu_path' in
settings:`.
- Fixes: #31
- Fixes: #32
- Fixes inconsistent timeline scaling. The predict tab now also uses a 5
frames per pixel ratio.

---------

Co-authored-by: diamondax03 <axelvandenberg4@gmail.com>
Co-authored-by: oThommy <tvdhil@datho.com>